requirements.txt
EasyProcess==0.3
et-xmlfile==1.1.0
greenlet==1.1.1
MouseInfo==0.1.3
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
playwright==1.14.0
pyee==8.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
python-dotenv==0.19.0
PyTweening==1.0.3
pytz==2021.1
PyVirtualDisplay==2.2
six==1.16.0
urllib3==1.26.6
websockets==9.1
